Handle sqlalchemy-migrate for keystone and nova
Add a lib in devstack to install CockroachDB. This lib provides two
functions: `database_start_switch_cockroachdb` and
`database_stop_switch_cockroachdb`. The soft `stack.sh`, that tells
which database backend to use, is patched so, devstack installs both
postgresql and cockroachdb. Then we put switching function to tell a
service use the one or other.